Answered self-study questions and installed styling libraries
