+++ README.md
@@ -25,12 +25,12 @@ In this challenge, you will create a Single Page Application complete with Clien
 Demonstrate your understanding of this Sprint's concepts by answering the following free-form questions. Edit this document to include your answers after each question.
 
 - [ ] Explain benefit(s) using `client-side routing`?
-> Answer:
+> Answer: Smooth seamless transitions mostly used in single page applications (no refreshing)
 - [ ] What does AJAX stand for?
-> Answer:
+> Answer:AJAX stands for asynchronous Javascript and XML
 - [ ] What are `controlled components` in React?
 > Answer:
-
+When form data is handled by a React component. 
 
 ## Project Set Up
