+++ package.json
@@ -6,6 +6,7 @@
     "@material-ui/core": "^4.2.1",
     "axios": "^0.19.0",
     "react": "^16.8.6",
+    "react-bootstrap": "^1.0.0-beta.12",
     "react-dom": "^16.8.6",
     "react-router-dom": "^5.0.1",
     "react-scripts": "^3.0.1",
