+++ src/components/WelcomePage.js
@@ -7,7 +7,7 @@ export default function WelcomePage() {
         <h1>Welcome to the ultimate fan site!</h1>
         <img
           className="main-img ui centered medium circular image"
-          src="https://rickandmortyapi.com/api/character/avatar/1.jpeg"
+          src="https://cdn.pastemagazine.com/www/articles/1111.jpg"
           alt="rick"
         />
       </header>
