+++ src/index.js
@@ -4,6 +4,7 @@ import { BrowserRouter as Router } from "react-router-dom";
 import "semantic-ui-css/semantic.min.css";
 import "./index.css";
 import App from "./App";
+// import 'bootstrap/dist/css/bootstrap.min.css';
 
 ReactDOM.render(
   <Router>
